requirements.txt
fastapi
uvicorn[standard]
python-multipart
pandas
openpyxl
xlrd
pdfplumber
anthropic
python-dotenv
